feat(pipeline): unify analysis with layout-aware static extraction
Move orchestration into pipeline.analyze. When PE/ELF/Mach-O layout
succeeds, reuse quantum extract (tags, structures, FP-lib filter) and
replace static_strings; fall back to classic statics otherwise.
CLI gains --no layout/tags; main is a thin entry point.
